fix(docker): missing config file in dev image (#14753)
<!-- First of all, THANK YOU for your contribution. -->

## Changes Proposed:
When we create the dev image we do not have the .dockerdist file under
the etc folder. The ac-dev-server included in the azerothcore-wotlk
repository bind the /env/docker/etc folder inside the container, but the
acore-docker repository doesn't. That's why we are having the error in
that project.

This happened after the latest changes:

* https://github.com/azerothcore/azerothcore-wotlk/pull/14707
* https://github.com/azerothcore/azerothcore-wotlk/pull/14747

Closes https://github.com/azerothcore/acore-docker/issues/12